feat: add HookDecorator so a hook wrapper forwards the stages it does not override
A hook's stages each default to doing nothing, so a wrapper that omits one
swallows it rather than passing it on: wrapping a hook meant reimplementing
beforeIdentify, afterIdentify, and afterTrack as pure forwarding just to keep
them working, and a wrapper that forgot dropped them silently. HookDecorator
forwards every stage, so a subclass overrides only the stages it changes.
DedupingHook extends it and is down to the three it changes.

The wrapped hook is private to the decorator, reachable only through super, so
nothing can unwrap a decorator to decide what to do with what it finds. The
identify and track stages are @CallSuper, so lint reports an override that stops
forwarding one: a decorator wrapping a DedupingHook could otherwise keep it from
being told to forget what it has reported, and it would go on suppressing across
an identify. The evaluation stages are not, because suppressing an evaluation
series is what a decorator is for.

Also drops the remaining suggestions to subclass from the deduper docs, which the
iOS SDK had already removed.
