+++ launchdarkly-android-client-sdk/src/main/java/com/launchdarkly/sdk/android/integrations/DedupingHook.java
@@ -28,12 +28,11 @@ import java.util.Objects;
  *         .addHook(new MetricsHook())                                  // observes every evaluation
  *         .addHook(new DedupingHook(new ObservabilityHook()))           // default window
  *         .addHook(new DedupingHook(new TelemetryHook(), 60_000))
- *         .addHook(new DedupingHook(new ExperimentHook(), myCustomDeduper))
+ *         .addHook(new DedupingHook(new ExperimentHook(), sharedDeduper))
  * </code></pre>
  * <p>
  * Two evaluations resolve to the same result when they agree on everything
- * {@link EvaluationExposureKey} describes. Pass your own {@link EvaluationExposureDeduper} subclass to
- * decide that differently.
+ * {@link EvaluationExposureKey} describes.
  * <p>
  * An evaluation the SDK has no flag data for resolves to the default value, and is the same result as
  * another that does. Evaluations made before the client has flags are of that kind, so the wrapped
@@ -58,7 +57,7 @@ import java.util.Objects;
  * it stored in its own before stage. A decorator inside this one is unaffected, since a suppressed
  * evaluation never reaches it.
  */
-public final class DedupingHook extends Hook {
+public final class DedupingHook extends HookDecorator {
 
     /**
      * Reads the clock a window is measured against. Exists so that tests can control it; the SDK has
@@ -80,7 +79,6 @@ public final class DedupingHook extends Hook {
     // Namespaced because it travels in series data that the wrapped hook may also write to.
     private static final String SUPPRESSED = "com.launchdarkly.sdk.android.DedupingHook.suppressed";
 
-    private final Hook delegate;
     private final EvaluationExposureDeduper deduper;
     private final Clock clock;
 
@@ -117,18 +115,9 @@ public final class DedupingHook extends Hook {
 
     @VisibleForTesting
     DedupingHook(Hook delegate, EvaluationExposureDeduper deduper, Clock clock) {
-        super(nameOf(delegate));
-        this.delegate = Objects.requireNonNull(delegate, "a deduping hook must wrap a hook");
+        super(delegate);
         this.deduper = Objects.requireNonNull(deduper, "a deduping hook must have a deduper");
         this.clock = clock;
-    }
-
-    /**
-     * @return the wrapped hook's metadata, so that the SDK names the hook a stage belongs to
-     */
-    @Override
-    public HookMetadata getMetadata() {
-        return delegate.getMetadata();
     }
 
     /**
@@ -148,7 +137,7 @@ public final class DedupingHook extends Hook {
         if (key != null && !deduper.shouldRecord(key, clock.elapsedMillis())) {
             return suppressedSeriesData;
         }
-        return delegate.beforeEvaluation(seriesContext, seriesData);
+        return super.beforeEvaluation(seriesContext, seriesData);
     }
 
     /**
@@ -165,7 +154,7 @@ public final class DedupingHook extends Hook {
         if (seriesData != null && seriesData.get(SUPPRESSED) == this) {
             return seriesData;
         }
-        return delegate.afterEvaluation(seriesContext, seriesData, evaluationDetail);
+        return super.afterEvaluation(seriesContext, seriesData, evaluationDetail);
     }
 
     /**
@@ -183,27 +172,6 @@ public final class DedupingHook extends Hook {
     @Override
     public Map<String, Object> beforeIdentify(IdentifySeriesContext seriesContext, Map<String, Object> seriesData) {
         deduper.reset();
-        return delegate.beforeIdentify(seriesContext, seriesData);
-    }
-
-    @Override
-    public Map<String, Object> afterIdentify(IdentifySeriesContext seriesContext, Map<String, Object> seriesData,
-                                             IdentifySeriesResult result) {
-        return delegate.afterIdentify(seriesContext, seriesData, result);
-    }
-
-    @Override
-    public void afterTrack(TrackSeriesContext seriesContext) {
-        delegate.afterTrack(seriesContext);
-    }
-
-    // Static because it runs in the super() call, before this instance exists. Tolerates a hook whose
-    // metadata throws, which the SDK reports rather than propagates.
-    private static String nameOf(Hook delegate) {
-        try {
-            return delegate == null ? null : delegate.getMetadata().getName();
-        } catch (Exception e) {
-            return null;
-        }
+        return super.beforeIdentify(seriesContext, seriesData);
     }
 }
